refactor(scripts): Move slide-import-lib to src/jobs directory
- Moved scripts/slide-import-lib.ts to src/jobs/slide-import-lib.ts
- Updated import in src/app/api/cron/update-slides/route.ts
- Fixes TypeScript path resolution (files in scripts/ not included in tsconfig)
- Now uses @/jobs path alias which works correctly with tsc
