+++ frontend/src/jobs/slide-import-lib.ts
